move listeners to one place and add some docs
